chore: move buffer to evm_core from debugger
