+++ crates/cheatcodes/Cargo.toml
@@ -29,7 +29,6 @@ foundry-config.workspace = true
 foundry-evm-core.workspace = true
 foundry-evm-traces.workspace = true
 foundry-wallets.workspace = true
-foundry-debugger.workspace = true
 
 alloy-dyn-abi.workspace = true
 alloy-json-abi.workspace = true
