+++ crates/cheatcodes/src/evm/record_debug_step.rs
@@ -3,7 +3,7 @@ use alloy_primitives::U256;
 use foundry_evm_traces::CallTraceArena;
 use revm::interpreter::{InstructionResult, OpCode};
 
-use foundry_debugger::tui::{context::BufferKind, draw::get_buffer_accesses};
+use foundry_evm_core::buffer::{get_buffer_accesses, BufferKind};
 use revm_inspectors::tracing::types::{CallTraceStep, RecordedMemory};
 use spec::Vm::DebugStep;
 
@@ -40,11 +40,7 @@ pub(crate) fn convert_call_trace_to_debug_step(step: &CallTraceStep) -> DebugSte
     let opcode = step.op.get();
     let stack = get_stack_inputs_for_opcode(opcode, step.stack.as_ref());
 
-    let memory = get_memory_input_for_opcode(
-        opcode,
-        step.stack.as_ref(),
-        step.memory.as_ref(),
-    );
+    let memory = get_memory_input_for_opcode(opcode, step.stack.as_ref(), step.memory.as_ref());
 
     let is_out_of_gas = step.status == InstructionResult::OutOfGas ||
         step.status == InstructionResult::MemoryOOG ||
@@ -64,22 +60,20 @@ pub(crate) fn convert_call_trace_to_debug_step(step: &CallTraceStep) -> DebugSte
 
 // The expected `stack` here is from the trace stack, where the top of the stack
 // is the last value of the vector
-fn get_memory_input_for_opcode(opcode: u8, stack: Option<&Vec<U256>>, memory: Option<&RecordedMemory>) -> Vec<u8> {
-    let Some(stack_data) = stack else {
-        return vec![]
-    };
-    let Some(memory_data) = memory else {
-        return vec![]
-    };
+fn get_memory_input_for_opcode(
+    opcode: u8,
+    stack: Option<&Vec<U256>>,
+    memory: Option<&RecordedMemory>,
+) -> Vec<u8> {
+    let Some(stack_data) = stack else { return vec![] };
+    let Some(memory_data) = memory else { return vec![] };
 
     if let Some(accesses) = get_buffer_accesses(opcode, stack_data) {
         if let Some((kind, access)) = accesses.read {
             return match kind {
-                BufferKind::Memory => get_slice_from_memory(
-                    memory_data.as_bytes(),
-                    access.offset,
-                    access.len
-                ),
+                BufferKind::Memory => {
+                    get_slice_from_memory(memory_data.as_bytes(), access.offset, access.len)
+                }
                 _ => vec![],
             }
         }
@@ -96,14 +90,12 @@ fn get_stack_inputs_for_opcode(opcode: u8, stack: Option<&Vec<U256>>) -> Vec<U25
         return vec![]
     };
 
-    let Some(stack_data) = stack else {
-        return vec![]
-    };
+    let Some(stack_data) = stack else { return vec![] };
 
     let stack_input_size = op.inputs();
     let mut inputs = Vec::new();
     for i in 0..stack_input_size {
-        inputs.push(peak_stack(&stack_data, i.into()));
+        inputs.push(peak_stack(stack_data, i.into()));
     }
     inputs
 }
