+++ crates/debugger/Cargo.toml
@@ -16,6 +16,7 @@ workspace = true
 foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-evm-traces.workspace = true
+foundry-evm-core.workspace = true
 revm-inspectors.workspace = true
 
 alloy-primitives.workspace = true
