+++ crates/evm/core/src/lib.rs
@@ -20,6 +20,7 @@ pub mod abi {
 mod ic;
 
 pub mod backend;
+pub mod buffer;
 pub mod constants;
 pub mod decode;
 pub mod fork;
